# PERF: Series.str.split(expand=True) for pyarrow-backed strings